Extract Service url resolving to own class
